+++ server/app/crud/__init__.py
@@ -1,4 +1,5 @@
 from .base import CRUDBase
 from .chat_crud import chat_crud, get_chat_crud, CRUDChat
 from .course_crud import course_crud, get_course_crud, CRUDCourse
+from .event_crud import event_crud, get_event_crud, CRUDEvent
 from .files_crud import files_crud, get_files_crud, CRUDFiles
